+++ Changelog.md
@@ -26,6 +26,7 @@
 - Fix Expression conditions that are "truthy" in vtank will now return true, instead of having to specifically return 1.0 for expression conditions
 - Fix boolean operators now return the right-most executed operand
 - Fix quoted expression strings that had multiple backticks at the end, which would remove all ending backticks
+- Add expressions: ustadd[wobject], ustopen[], ustsalvage[]
 
 ## 0.2.5 (2021-10-06) [UtilityBeltInstaller-0.2.5.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/dd7285b144ef9327bf8041c64841c22f/UtilityBeltInstaller-0.2.5.exe)
 - Add expressions: wobjectgetdoubleprop[], wobjectgetboolprop[], wobjectgetstringprop[], wobjecthasdata[], wobjectrequestdata[]